src/main/java/hu/benkoa/cs/Application.java
package hu.benkoa.cs;

import hu.benkoa.cs.log.LogReader;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.CommandLineRunner;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.context.ApplicationContext;
import org.springframework.context.annotation.Bean;

@SpringBootApplication
public class Application {

    @Autowired
    LogReader logReader;

    public static void main(String[] args) {
        SpringApplication.run(Application.class, args);
    }

    @Bean
    public CommandLineRunner commandLineRunner(ApplicationContext ctx) {
        return args -> {
            if (args.length>0) {
                this.logReader.processFile(args[0]);
            } else {
                throw new RuntimeException("Please provide path to log file as a command line argument");
            }
        };
    }

}
